requirements.txt
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.1
google-generativeai==0.5.4
